pyproject.toml
[project]
name = "gitreporag"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rmc8", email = "k@rmc-8.com" }
]
dependencies = [
    "langchain>=0.3.14",
    "langchain-core>=0.3.29",
    "gitpython>=3.1.44",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.14",
    "langchain-chroma>=0.1.4",
    "langchain-ollama>=0.2.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gitreporag"]
